moves HyperdriveAgent up
