+++ lib/agent0/agent0/hyperdrive/__init__.py
@@ -1,3 +1,4 @@
+from .agents.hyperdrive_account import HyperdriveAgent
 from .interface.read_interface import HyperdriveReadInterface
 from .interface.read_write_interface import HyperdriveReadWriteInterface
 from .policies.hyperdrive_policy import HyperdriveBasePolicy
