+++ lib/agent0/agent0/hyperdrive/agents/__init__.py
@@ -1,3 +1,1 @@
 """Account and wallet with Hyperdrive specific parts"""
-
-from .hyperdrive_account import HyperdriveAgent
