+++ lib/agent0/agent0/hyperdrive/exec/setup_experiment.py
@@ -7,8 +7,7 @@ from hyperlogs import setup_logging
 
 from agent0 import AccountKeyConfig
 from agent0.base.config import AgentConfig, EnvironmentConfig
-from agent0.hyperdrive import HyperdriveReadInterface
-from agent0.hyperdrive.agents import HyperdriveAgent
+from agent0.hyperdrive import HyperdriveAgent, HyperdriveReadInterface
 from agent0.hyperdrive.crash_report import setup_hyperdrive_crash_report_logging
 
 from .get_agent_accounts import get_agent_accounts
