+++ lib/agent0/bin/fund_agents_from_user_key.py
@@ -5,12 +5,13 @@ import asyncio
 import logging
 import os
 
-from agent0 import build_account_config_from_env
-from agent0.hyperdrive.agents import HyperdriveAgent
-from agent0.hyperdrive.exec import async_fund_agents
 from eth_account.account import Account
 from ethpy import build_eth_config
 from ethpy.hyperdrive import fetch_hyperdrive_address_from_uri
+
+from agent0 import build_account_config_from_env
+from agent0.hyperdrive import HyperdriveAgent
+from agent0.hyperdrive.exec import async_fund_agents
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(
